% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granova-package.R
\docType{package}
\name{granova-package}
\alias{granova-package}
\alias{granova}
\title{granova: Graphical Analysis of Variance}
\description{
This small collection of functions provides what we call elemental graphics for display of analysis of variance results, David C. Hoaglin, Frederick Mosteller and John W. Tukey (1991, ISBN:978-0-471-52735-0), Paul R. Rosenbaum (1989) \doi{10.2307/2684513}, Robert M. Pruzek and James E. Helmreich \url{https://jse.amstat.org/v17n1/helmreich.html}. The term elemental derives from the fact that each function is aimed at construction of graphical displays that afford direct visualizations of data with respect to the fundamental questions that drive the particular analysis of variance methods. These functions can be particularly helpful for students and non-statistician analysts. But these methods should be quite generally helpful for work-a-day applications of all kinds, as they can help to identify outliers, clusters or patterns, as well as highlight the role of non-linear transformations of data.
}
\references{
David C. Hoaglin, Frederick Mosteller and John W. Tukey 
(1991, ISBN:978-0-471-52735-0).

Paul R. Rosenbaum (1989)  <doi:10.2307/2684513>.

Robert M. Pruzek and James E. Helmreich <https://jse.amstat.org/v17n1/helmreich.html>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fbertran.github.io/granova/}
  \item \url{https://github.com/fbertran/granova/}
  \item Report bugs at \url{https://github.com/fbertran/granova/issues/}
}

}
\author{
This package has been written by Robert M. Pruzek and 
James E. Helmreich. Frederic Bertrand is now the maintainer of package and
has been doing several code fixes and improvements to the package.

Maintainer: Frederic Bertrand <frederic.bertrand@lecnam.net>
}
\keyword{internal}
